[C] rename waitlist->steamder
