+++ packages/client/public/locales/de/pages/steamder.json
@@ -1,5 +1,5 @@
 {
-    "waitlist": {
+    "steamder": {
         "header": "Du bist in einer Steamder, aber noch im Wartezimmer! Hier siehst du, mit wem du potenziell spielen wirst. Wir warten nur noch darauf, dass der Administrator (das Profil mit der kleinen Krone) den Startknopf drückt, dann kannst du anfangen auszuwählen!",
         "players": {
             "games_in_common_zero": "Keine Spiele gemeinsam... :(",
