+++ packages/client/public/locales/en/pages/steamder.json
@@ -1,5 +1,5 @@
 {
-    "waitlist": {
+    "steamder": {
         "header": "You're in a Steamder, but still in the waiting room! Here are your potential gaming buddies. We're just waiting for the admin (the profile with the little crown) to start, then you can begin picking games!",
         "players": {
             "games_in_common_zero": "No games in common... :(",
