+++ packages/client/public/locales/es/pages/steamder.json
@@ -1,5 +1,5 @@
 {
-    "waitlist": {
+    "steamder": {
         "header": "Estás en un Steamder, ¡pero aún en la sala de espera! Aquí están los jugadores con los que potencialmente jugarás, solo esperamos que el administrador del salón (el perfil con la pequeña corona) inicie, ¡y podrás comenzar a elegir!",
         "players": {
             "games_in_common_zero": "Ningún juego en común... :(",
