+++ packages/client/public/locales/fr/pages/steamder.json
@@ -1,5 +1,5 @@
 {
-    "waitlist": {
+    "steamder": {
         "header": "Vous êtes dans une Steamder, mais encore en salle d'attente ! Voici avec qui vous allez potentiellement jouer, on n'attends plus que l'administrateur du salon (le profil avec la petite couronne) démarre, et vous pourrez commencer à choisir !",
         "players": {
             "games_in_common_zero": "Aucun jeu en commun ... :(",
