+++ packages/client/public/locales/jp/pages/steamder.json
@@ -1,5 +1,5 @@
 {
-    "waitlist": {
+    "steamder": {
         "header": "Steamderの待合室にいます！これから一緒に遊ぶかもしれない人たちです。管理者（小さな王冠のついたプロファイル）が開始するのを待って、ゲーム選択を始められます！",
         "players": {
             "games_in_common_zero": "共通のゲームがありません... :(",
