+++ packages/client/src/components/organisms/homepage/StatsHome.tsx
@@ -94,7 +94,7 @@ export const StatsHome: React.FC = () => {
                             <h2 className="fw-bold mb-0">
                                 { isPending && <Spinner animation="border" /> }
                                 { !isPending &&
-                                    <CountUp separator=" " start={ 0 } end={ data.waitlists } duration={ 5 } enableScrollSpy={true} scrollSpyOnce={true} scrollSpyDelay={250}>
+                                    <CountUp separator=" " start={ 0 } end={ data.steamders } duration={ 5 } enableScrollSpy={true} scrollSpyOnce={true} scrollSpyDelay={250}>
                                         {({ countUpRef }) => <span ref={countUpRef} />}
                                     </CountUp>
                                 }
