+++ packages/client/src/types/StatsApiResponse.ts
@@ -2,7 +2,7 @@ export interface StatsApiResponse {
     data: {
       players: number;
       games: number;
-      waitlists: number;
+      steamders: number;
       matches: number;
     };
     message: string;
